# runpod-serverless/openapi.yaml
openapi: 3.0.3
info:
  title: Rock/Pop Music Generator API
  description: API for generating AI-driven rock and pop music with lyrics, style tags, and multi-track MIDI extraction using HeartMuLa, YourMT3+, and Basic Pitch.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://{pod_id}-8000.proxy.runpod.net
    description: RunPod Proxy URL (example)
  - url: https://api.runpod.ai/v1/{endpoint_id}/runsync
    description: RunPod Serverless Sync API
  - url: https://api.runpod.ai/v1/{endpoint_id}/run
    description: RunPod Serverless Async API
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    GenerateRequest:
      type: object
      required:
        - lyrics
      properties:
        lyrics:
          type: string
          description: The lyrics for the music generation.
          example: "I'm feeling good, it's a sunny day..."
        style:
          type: string
          description: Style tags for the generation.
          default: "indie rock, electric guitar, driving drums, male vocal, E minor"
          example: "heavy metal, distorted guitar, fast drums"
        duration:
          type: integer
          description: Target duration in seconds.
          default: 180
          example: 120
        seed:
          type: integer
          description: Random seed for reproducible generation. -1 for random.
          default: -1
          example: 42
        stems:
          type: boolean
          description: Whether to generate individual stems (vocals, drums, etc.)
          default: false
        midi:
          type: boolean
          description: Whether to generate MIDI tracks from the stems or audio.
          default: false
        midi_model:
          type: string
          description: The model to use for MIDI extraction.
          enum: [pitch, mt3]
          default: pitch
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, generating_audio, extracting_midi, completed, failed, not_found]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        audio_url:
          type: string
          description: URL to download the main audio file (WAV).
          nullable: true
        midi_url:
          type: string
          description: URL to download the MIDI file. Only present if stems=true and midi=true (per stem).
          nullable: true
        zip_url:
          type: string
          description: URL to download a ZIP archive containing all generated assets (stems, MIDI, etc).
          nullable: true
        error:
          type: string
          nullable: true
paths:
  /generate:
    post:
      summary: Start a music generation task
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Task accepted and processing in background
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: "processing"
        '401':
          description: Unauthorized

  /task/{task_id}:
    get:
      summary: Get the status and results of a generation task
      security:
        - basicAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '401':
          description: Unauthorized

  /generations:
    get:
      summary: List all generation tasks for the authenticated user
      security:
        - basicAuth: []
      responses:
        '200':
          description: List of generations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: string
                      format: uuid
                    username:
                      type: string
                    status:
                      type: string
                    created_at:
                      type: string
                    downloads:
                      type: object
                      properties:
                        audio:
                          type: string
                        zip:
                          type: string
                        metadata:
                          type: string
        '401':
          description: Unauthorized

  /download/{username}/{task_id}/{filename}:
    get:
      summary: Download generated audio, MIDI, or ZIP files
      security:
        - basicAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: "The name of the file to download. Examples: {task_id}.wav, {task_id}.mid, {task_id}.zip, metadata.json"
          example: "550e8400-e29b-41d4-a716-446655440000.zip"
      responses:
        '200':
          description: The file content
          content:
            audio/wav:
              schema:
                type: string
                format: binary
            audio/midi:
              schema:
                type: string
                format: binary
            application/zip:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (trying to download another user's files)
        '404':
          description: File not found

  /ws/{task_id}:
    get:
      summary: WebSocket for real-time task updates
      description: Subscribe to task updates via WebSocket. Receives the same payload as the `/task/{task_id}` endpoint whenever progress changes.
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '101':
          description: Switching protocols (WebSocket upgrade)